Conflicts:
	bindings/python/mapnik_text_placement.cpp
	include/mapnik/font_engine_freetype.hpp
	include/mapnik/map.hpp
	include/mapnik/metawriter.hpp
	include/mapnik/metawriter_inmem.hpp
	include/mapnik/metawriter_json.hpp
	include/mapnik/placement_finder.hpp
	include/mapnik/symbolizer.hpp
	include/mapnik/symbolizer_helpers.hpp
	include/mapnik/text_properties.hpp
	src/agg/process_shield_symbolizer.cpp
	src/build.py
	src/cairo_renderer.cpp
	src/feature_style_processor.cpp
	src/grid/process_shield_symbolizer.cpp
	src/metawriter.cpp
	src/metawriter_inmem.cpp
	src/placement_finder.cpp
	src/processed_text.cpp
	src/symbolizer_helpers.cpp
	tests/visual_tests/images/lines-shield-200-reference.png
	tests/visual_tests/images/lines-shield-400-reference.png
	tests/visual_tests/images/lines-shield-600-reference.png
	tests/visual_tests/images/lines-shield-800-reference.png
	tests/visual_tests/styles/formatting.xml
	tests/visual_tests/test.py

+++ include/mapnik/text/face.hpp
@@ -67,7 +67,7 @@ public:
 
     double get_char_height() const;
 
-    bool set_character_sizes(float size);
+    bool set_character_sizes(double size);
 
     void glyph_dimensions(glyph_info &glyph) const;
 
@@ -88,7 +88,7 @@ public:
     font_face_set(void) : faces_(){}
 
     void add(face_ptr face);
-    void set_character_sizes(float size);
+    void set_character_sizes(double size);
 
     unsigned size() const { return faces_.size(); }
     iterator begin() { return faces_.begin(); }
